# docs/swagger.yaml
definitions:
...
  requestmodel.Order:
    properties:
      addressid:
        type: string
      couponid:
        type: string
      paymentMethod:
        type: string
      userid:
        type: string
    required:
    - addressid
    - paymentMethod
    - userid
    type: object
...
  response.Response:
    properties:
      error: {}
      message:
        type: string
      result: {}
      status_code:
        type: integer
    type: object
...
info:
  contact:
    email: shahabazsulthan4@gmail.com
    name: 'Creator: Shahabaz Sulthan'
  description: "Laptop Lounge - \U0001F680 Your One-Stop Destination for Ultimate
    Laptop Shopping!\n\U0001F6D2\U0001F4BB Browse, compare, and buy top-notch laptops
    effortlessly.\nPowered by cutting-edge technology, we bring you a seamless shopping
    experience.\n\U0001F31F Dive into the future of laptop shopping with Laptop Lounge!
    \U0001F680\U0001F525"
  title: Laptop Lounge API
paths:
...
  /invoice/{orderItemID}:
    get:
      consumes:
      - application/json
      description: Generate an invoice for the specified order item ID
      parameters:
      - description: Order Item ID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - RefreshToken: []
      summary: Generate invoice
      tags:
      - User Invoice
...
  /order:
    get:
      consumes:
      - application/json
      description: Get a list of all orders for the logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - Refreshtoken: []
      summary: Get all orders
      tags:
      - User Orders
    post:
      consumes:
      - application/json
      description: Create a new order with the input payload
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/requestmodel.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - Refreshtoken: []
      summary: Create a new order
      tags:
      - User Orders
  /order/{orderItemID}:
    get:
      consumes:
      - application/json
      description: Get the details of a single order by its ID
      parameters:
      - description: Order Item ID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - Refreshtoken: []
      summary: Get order details
      tags:
      - User Orders
    patch:
      consumes:
      - application/json
      description: Cancel an order by its ID for the logged-in user
      parameters:
      - description: Order Item ID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - RefreshToken: []
      summary: Cancel order
      tags:
      - User Orders
  /order/return/{orderItemID}:
    patch:
      consumes:
      - application/json
      description: Return an order by its ID for the logged-in user
      parameters:
      - description: Order Item ID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - RefreshToken: []
      summary: Return order
      tags:
      - User Orders
...
  /seller/order/{SellerID}:
    get:
      consumes:
      - application/json
      description: Get all orders for the specified seller ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all seller orders
      tags:
      - Order Management
  /seller/order/{SellerID}/{orderItemID}:
    patch:
      consumes:
      - application/json
      description: Confirm an order as delivered by its ID for the specified seller
        ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: Order Item ID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Confirm order delivered
      tags:
      - Order Management
  /seller/order/cancel/{SellerID}/{orderID}:
    patch:
      consumes:
      - application/json
      description: Cancel an order by its ID for the specified seller ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Cancel order
      tags:
      - Order Management
  /seller/order/cancelled/{SellerID}:
    get:
      consumes:
      - application/json
      description: Get all cancelled orders for the specified seller ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get seller cancelled orders
      tags:
      - Order Management
...
  /seller/order/processing/{SellerID}:
    get:
      consumes:
      - application/json
      description: Get all processing orders for the specified seller
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get seller processing orders
      tags:
      - Order Management
...
  /seller/report/day/{SellerID}/{year}/{month}/{day}:
    get:
      consumes:
      - application/json
      description: Generate a sales report for the specified seller ID for the past
        number of days
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: string
      - description: Month
        in: query
        name: month
        required: true
        type: string
      - description: Day
        in: query
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Generate sales report for custom days
      tags:
      - Seller Report
  /seller/report/days/{SellerID}/{days}:
    get:
      consumes:
      - application/json
      description: Generate a custom sales report for the specified seller ID for
        the given number of days
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      - description: Number of days
        in: path
        name: days
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Generate custom sales report
      tags:
      - Seller Report
  /seller/report/pdf/{SellerID}:
    get:
      consumes:
      - application/json
      description: Generate a sales report PDF for the specified seller ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales report created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Seller not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Generate sales report PDF
      tags:
      - Seller Report
  /seller/report/xlsx/{SellerID}:
    get:
      consumes:
      - application/json
      description: Generate a sales report in XLSX format for the specified seller
        ID
      parameters:
      - description: Seller ID
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Generate sales report in XLSX
      tags:
      - Seller Report
...
securityDefinitions:
  BearerTokenAuth:
    description: 'Use your Bearer token for authentication. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Use your Refresh token to obtain a new Bearer token. Example: "{token}"'
    in: header
    name: Refreshtoken
    type: apiKey
swagger: "2.0"